+++ yolort/runtime/trt_helper.py
@@ -11,16 +11,124 @@
 
 import logging
 from pathlib import Path
-from typing import Optional
+from typing import Optional, Tuple, Union
 
 try:
     import tensorrt as trt
 except ImportError:
     trt = None
 
+import torch
+from torch import nn, Tensor
+from yolort.models import YOLO
+from yolort.models.anchor_utils import AnchorGenerator
+from yolort.models.backbone_utils import darknet_pan_backbone
+from yolort.models.box_head import LogitsDecoder
+from yolort.utils import load_from_ultralytics
+
 logging.basicConfig(level=logging.INFO)
-logging.getLogger("EngineBuilder").setLevel(logging.INFO)
-log = logging.getLogger("EngineBuilder")
+logging.getLogger("TRTHelper").setLevel(logging.INFO)
+logger = logging.getLogger("TRTHelper")
+
+
+class YOLOTRTModule(nn.Module):
+    """
+    TensorRT deployment friendly wrapper for YOLO.
+
+    Remove the ``torchvision::nms`` in this warpper, due to the fact that some third-party
+    inference frameworks currently do not support this operator very well.
+    """
+
+    def __init__(
+        self,
+        checkpoint_path: str,
+        version: str = "r6.0",
+    ):
+        super().__init__()
+        model_info = load_from_ultralytics(checkpoint_path, version=version)
+
+        backbone_name = f"darknet_{model_info['size']}_{version.replace('.', '_')}"
+        depth_multiple = model_info["depth_multiple"]
+        width_multiple = model_info["width_multiple"]
+        use_p6 = model_info["use_p6"]
+        backbone = darknet_pan_backbone(
+            backbone_name,
+            depth_multiple,
+            width_multiple,
+            version=version,
+            use_p6=use_p6,
+        )
+        num_classes = model_info["num_classes"]
+        anchor_generator = AnchorGenerator(model_info["strides"], model_info["anchor_grids"])
+        post_process = LogitsDecoder(model_info["strides"])
+        model = YOLO(
+            backbone,
+            num_classes,
+            anchor_generator=anchor_generator,
+            post_process=post_process,
+        )
+
+        model.load_state_dict(model_info["state_dict"])
+        self.model = model
+        self.num_clases = num_classes
+
+    @torch.no_grad()
+    def forward(self, inputs: Tensor) -> Tuple[Tensor, Tensor]:
+        """
+        Args:
+            inputs (Tensor): batched images, of shape [batch_size x 3 x H x W]
+        """
+        # Compute the detections
+        outputs = self.model(inputs)
+
+        return outputs
+
+    @torch.no_grad()
+    def to_onnx(
+        self,
+        file_path: Union[str, Path],
+        input_sample: Optional[Tensor] = None,
+        opset_version: int = 11,
+        enable_dynamic: bool = True,
+        **kwargs,
+    ):
+        """
+        Saves the model in ONNX format.
+
+        Args:
+            file_path: The path of the file the onnx model should be saved to.
+            input_sample: An input for tracing. Default: None.
+            opset_version: Opset version we export the model to the onnx submodule. Default: 11.
+            enable_dynamic: Whether to specify axes of tensors as dynamic. Default: True.
+            **kwargs: Will be passed to torch.onnx.export function.
+        """
+        if input_sample is None:
+            input_sample = torch.rand(1, 3, 320, 320).to(next(self.parameters()).device)
+
+        dynamic_axes = (
+            {
+                "images": {0: "batch", 2: "height", 3: "width"},
+                "boxes": {0: "batch", 1: "num_objects"},
+                "scores": {0: "batch", 1: "num_objects"},
+            }
+            if enable_dynamic
+            else None
+        )
+
+        input_names = ["images"]
+        output_names = ["boxes", "scores"]
+
+        torch.onnx.export(
+            self.model,
+            input_sample,
+            file_path,
+            do_constant_folding=True,
+            opset_version=opset_version,
+            input_names=input_names,
+            output_names=output_names,
+            dynamic_axes=dynamic_axes,
+            **kwargs,
+        )
 
 
 class EngineBuilder:
@@ -66,12 +174,12 @@ class EngineBuilder:
         inputs = [self.network.get_input(i) for i in range(self.network.num_inputs)]
         outputs = [self.network.get_output(i) for i in range(self.network.num_outputs)]
 
-        log.info("Network Description")
+        logger.info("Network Description")
         for input in inputs:
             self.batch_size = input.shape[0]
-            log.info(f"Input '{input.name}' with shape {input.shape} and dtype {input.dtype}")
+            logger.info(f"Input '{input.name}' with shape {input.shape} and dtype {input.dtype}")
         for output in outputs:
-            log.info(f"Output '{output.name}' with shape {output.shape} and dtype {output.dtype}")
+            logger.info(f"Output '{output.name}' with shape {output.shape} and dtype {output.dtype}")
         assert self.batch_size > 0
         self.builder.max_batch_size = self.batch_size
 
@@ -100,18 +208,18 @@ class EngineBuilder:
 
         engine_path.parent.mkdir(parents=True, exist_ok=True)
 
-        log.info(f"Building {precision} Engine in {engine_path}")
+        logger.info(f"Building {precision} Engine in {engine_path}")
 
         if precision == "fp16":
             if not self.builder.platform_has_fast_fp16:
-                log.warning("FP16 is not supported natively on this platform/device")
+                logger.warning("FP16 is not supported natively on this platform/device")
             else:
                 self.config.set_flag(trt.BuilderFlag.FP16)
         elif precision == "fp32":
-            log.info("Using fp32 mode.")
+            logger.info("Using fp32 mode.")
         else:
             raise NotImplementedError(f"Currently hasn't been implemented: {precision}.")
 
         with self.builder.build_engine(self.network, self.config) as engine, open(engine_path, "wb") as f:
             f.write(engine.serialize())
-            log.info(f"Serialize engine success, saved as {engine_path}")
+            logger.info(f"Serialize engine success, saved as {engine_path}")
